Fixed bug in Double.parseDouble() (introduced by me earlier). Refined Juan's code in NMEASentence some more. Also changed GPS classes to use doubles instead of floats for lat/long.
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@2388 06c8f799-c779-4783-8af7-179afedb67c8

+++ classes/lejos/gps/SimpleGPS.java
@@ -71,7 +71,7 @@ public class SimpleGPS extends Thread {
 	 * 
 	 * @return the latitude
 	 */
-	public float getLatitude() {
+	public double getLatitude() {
 		return ggaSentence.getLatitude();
 	}
 
@@ -91,7 +91,7 @@ public class SimpleGPS extends Thread {
 	 * 
 	 * @return the longitude
 	 */
-	public float getLongitude() {
+	public double getLongitude() {
 		return ggaSentence.getLongitude();
 	}
